// nft-example-client/src/types/token.types.ts
// Generated by https://quicktype.io

export interface INFTToken {
  token_address: string;
  token_id: string;
  amount: string;
  contract_type: string;
  name: null;
  symbol: null;
  token_uri: string;
  metadata: Record<string, string>;
  synced_at: string;
  owners?: string[];
}

// Generated by https://quicktype.io

export interface ITokenIdOwner {
  token_address: string;
  token_id: string;
  block_number_minted: string;
  owner_of: string;
  block_number: string;
  amount: string;
  contract_type: string;
  name: null;
  symbol: null;
  token_uri: string;
  metadata: string;
  synced_at: string;
  is_valid: number;
  syncing: number;
  frozen: number;
}